Change the way analog signal is converted

+++ ClientWpfApp/ServiceClients/ModbusServiceClient.cs
@@ -35,7 +35,7 @@ namespace ClientWpfApp.ServiceClients
 			});
 		}
 
-		public void WriteAnalogSignalValue(int rtuId, int signalAddress, int value)
+		public void WriteAnalogSignalValue(int rtuId, int signalAddress, double value)
 		{
 			modbusDuplexClient.WriteAnalogSignal(rtuId, signalAddress, value);
 		}
